requirements.txt
Django==1.10
djangorestframework==3.4.1
